+++ pyodide_build/tests/test_pywasmcross.py
@@ -257,7 +257,6 @@ def test_get_cmake_compiler_flags():
 
 def test_handle_command_cmake(build_args):
     args = build_args
-    assert "--fresh" in handle_command_generate_args(["cmake", "./"], args)
 
     build_cmd = ["cmake", "--build", "." "--target", "target"]
     assert handle_command_generate_args(build_cmd, args) == build_cmd